+++ Makefile
@@ -17,8 +17,9 @@ computing-provider:
 	$(GOCC) build $(GOFLAGS) -o computing-provider ./cmd/computing-provider
 .PHONY: computing-provider
 
-install:
+install: computing-provider
 	sudo install -C computing-provider /usr/local/bin/computing-provider
+	export CP_PATH=".swan_node"
 
 clean:
 	sudo rm -rf /usr/local/bin/computing-provider
