+++ internal/computing/cp_service.go
@@ -7,26 +7,26 @@ import (
 	"fmt"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/ethereum/go-ethereum/ethclient"
+	// "github.com/ethereum/go-ethereum/ethclient"
 	"github.com/filswan/go-mcs-sdk/mcs/api/common/logs"
 	"github.com/gin-gonic/gin"
 	"github.com/gomodule/redigo/redis"
 	"github.com/google/uuid"
 	"github.com/gorilla/websocket"
-	"github.com/joho/godotenv"
-	"github.com/swanchain/go-computing-provider/account"
+	// "github.com/joho/godotenv"
+	// "github.com/swanchain/go-computing-provider/account"
 	"github.com/swanchain/go-computing-provider/build"
 	"github.com/swanchain/go-computing-provider/conf"
 	"github.com/swanchain/go-computing-provider/constants"
 	"github.com/swanchain/go-computing-provider/internal/models"
 	"github.com/swanchain/go-computing-provider/util"
-	"github.com/swanchain/go-computing-provider/wallet"
+	// "github.com/swanchain/go-computing-provider/wallet"
 	"io"
 	batchv1 "k8s.io/api/batch/v1"
-	coreV1 "k8s.io/api/core/v1"
+	// coreV1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
+	// "k8s.io/apimachinery/pkg/api/resource"
 	metaV1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"math/rand"
@@ -596,365 +596,27 @@ func DoUbiTask(c *gin.Context) {
 		return
 	}
 
-	if strings.TrimSpace(ubiTask.InputParam) == "" {
-		c.JSON(http.StatusBadRequest, util.CreateErrorResponse(util.UbiTaskParamError, "missing required field: input_param"))
-		return
+	if constants.GetUBIType(ubiTask.ZkType) == constants.FIL_C2 {
+		Fil_C2_DoUbiTask(c, ubiTask)
+	} else if constants.GetUBIType(ubiTask.ZkType) == constants.ALEO_PROVER {
+		Aleo_Prover_DoUbiTask(c, ubiTask)
 	}
-
-	if strings.TrimSpace(ubiTask.Signature) == "" {
-		c.JSON(http.StatusBadRequest, util.CreateErrorResponse(util.UbiTaskParamError, "missing required field: signature"))
-		return
-	}
-
-	cpRepoPath, _ := os.LookupEnv("CP_PATH")
-	nodeID := GetNodeId(cpRepoPath)
-
-	signature, err := verifySignature(conf.GetConfig().UBI.UbiEnginePk, fmt.Sprintf("%s%d", nodeID, ubiTask.ID), ubiTask.Signature)
-	if err != nil {
-		logs.GetLogger().Errorf("verifySignature for ubi task failed, error: %+v", err)
-		c.JSON(http.StatusInternalServerError, util.CreateErrorResponse(util.UbiTaskParamError, "sign data failed"))
-		return
-	}
-
-	logs.GetLogger().Infof("ubi task sign verifing, task_id: %d, type: %s, verify: %v", ubiTask.ID, ubiTask.ZkType, signature)
-	if !signature {
-		c.JSON(http.StatusInternalServerError, util.CreateErrorResponse(util.UbiTaskParamError, "signature verify failed"))
-		return
-	}
-
-	var gpuFlag = "0"
-	var ubiTaskToRedis = new(models.CacheUbiTaskDetail)
-	ubiTaskToRedis.TaskId = strconv.Itoa(ubiTask.ID)
-	ubiTaskToRedis.TaskType = "CPU"
-	if ubiTask.Type == 1 {
-		ubiTaskToRedis.TaskType = "GPU"
-		gpuFlag = "1"
-	}
-	ubiTaskToRedis.Status = constants.UBI_TASK_RECEIVED_STATUS
-	ubiTaskToRedis.ZkType = ubiTask.ZkType
-	ubiTaskToRedis.CreateTime = time.Now().Format("2006-01-02 15:04:05")
-	SaveUbiTaskMetadata(ubiTaskToRedis)
-
-	var envFilePath string
-	envFilePath = filepath.Join(os.Getenv("CP_PATH"), "fil-c2.env")
-	envVars, err := godotenv.Read(envFilePath)
-	if err != nil {
-		logs.GetLogger().Errorf("reading fil-c2-env.env failed, error: %v", err)
-		return
-	}
-
-	c2GpuConfig := envVars["RUST_GPU_TOOLS_CUSTOM_GPU"]
-	c2GpuConfig = convertGpuName(strings.TrimSpace(c2GpuConfig))
-	nodeName, architecture, needCpu, needMemory, needStorage, err := checkResourceAvailableForUbi(ubiTask.Type, c2GpuConfig, ubiTask.Resource)
-	if err != nil {
-		ubiTaskToRedis.Status = constants.UBI_TASK_FAILED_STATUS
-		SaveUbiTaskMetadata(ubiTaskToRedis)
-		logs.GetLogger().Errorf("check resource failed, error: %v", err)
-		c.JSON(http.StatusInternalServerError, util.CreateErrorResponse(util.CheckResourcesError))
-		return
-	}
-
-	if nodeName == "" {
-		ubiTaskToRedis.Status = constants.UBI_TASK_FAILED_STATUS
-		SaveUbiTaskMetadata(ubiTaskToRedis)
-		logs.GetLogger().Warnf("ubi task id: %d, type: %s, not found a resources available", ubiTask.ID, ubiTaskToRedis.TaskType)
-		c.JSON(http.StatusInternalServerError, util.CreateErrorResponse(util.CheckAvailableResources))
-		return
-	}
-
-	var ubiTaskImage string
-	if architecture == constants.CPU_AMD {
-		ubiTaskImage = build.UBITaskImageAmdCpu
-		if gpuFlag == "1" {
-			ubiTaskImage = build.UBITaskImageAmdGpu
-		}
-	} else if architecture == constants.CPU_INTEL {
-		ubiTaskImage = build.UBITaskImageIntelCpu
-		if gpuFlag == "1" {
-			ubiTaskImage = build.UBITaskImageIntelGpu
-		}
-	}
-
-	mem := strings.Split(strings.TrimSpace(ubiTask.Resource.Memory), " ")[1]
-	memUnit := strings.ReplaceAll(mem, "B", "")
-	disk := strings.Split(strings.TrimSpace(ubiTask.Resource.Storage), " ")[1]
-	diskUnit := strings.ReplaceAll(disk, "B", "")
-	memQuantity, err := resource.ParseQuantity(fmt.Sprintf("%d%s", needMemory, memUnit))
-	if err != nil {
-		ubiTaskToRedis.Status = constants.UBI_TASK_FAILED_STATUS
-		SaveUbiTaskMetadata(ubiTaskToRedis)
-		logs.GetLogger().Error("get memory failed, error: %+v", err)
-		return
-	}
-
-	storageQuantity, err := resource.ParseQuantity(fmt.Sprintf("%d%s", needStorage, diskUnit))
-	if err != nil {
-		ubiTaskToRedis.Status = constants.UBI_TASK_FAILED_STATUS
-		SaveUbiTaskMetadata(ubiTaskToRedis)
-		logs.GetLogger().Error("get storage failed, error: %+v", err)
-		return
-	}
-
-	maxMemQuantity, err := resource.ParseQuantity(fmt.Sprintf("%d%s", needMemory*2, memUnit))
-	if err != nil {
-		ubiTaskToRedis.Status = constants.UBI_TASK_FAILED_STATUS
-		SaveUbiTaskMetadata(ubiTaskToRedis)
-		logs.GetLogger().Error("get memory failed, error: %+v", err)
-		return
-	}
-
-	maxStorageQuantity, err := resource.ParseQuantity(fmt.Sprintf("%d%s", needStorage*2, diskUnit))
-	if err != nil {
-		ubiTaskToRedis.Status = constants.UBI_TASK_FAILED_STATUS
-		SaveUbiTaskMetadata(ubiTaskToRedis)
-		logs.GetLogger().Error("get storage failed, error: %+v", err)
-		return
-	}
-
-	resourceRequirements := coreV1.ResourceRequirements{
-		Limits: coreV1.ResourceList{
-			coreV1.ResourceCPU:              *resource.NewQuantity(needCpu*2, resource.DecimalSI),
-			coreV1.ResourceMemory:           maxMemQuantity,
-			coreV1.ResourceEphemeralStorage: maxStorageQuantity,
-			"nvidia.com/gpu":                resource.MustParse(gpuFlag),
-		},
-		Requests: coreV1.ResourceList{
-			coreV1.ResourceCPU:              *resource.NewQuantity(needCpu, resource.DecimalSI),
-			coreV1.ResourceMemory:           memQuantity,
-			coreV1.ResourceEphemeralStorage: storageQuantity,
-			"nvidia.com/gpu":                resource.MustParse(gpuFlag),
-		},
-	}
-
-	go func() {
-		var namespace = "ubi-task-" + strconv.Itoa(ubiTask.ID)
-		var err error
-		defer func() {
-			key := constants.REDIS_UBI_C2_PERFIX + strconv.Itoa(ubiTask.ID)
-			ubiTaskRun, _ := RetrieveUbiTaskMetadata(key)
-			if ubiTaskRun.TaskId == "" {
-				ubiTaskRun = new(models.CacheUbiTaskDetail)
-				ubiTaskRun.TaskId = ubiTaskToRedis.TaskId
-				ubiTaskRun.TaskType = ubiTaskToRedis.TaskType
-				ubiTaskRun.ZkType = ubiTask.ZkType
-				ubiTaskRun.CreateTime = ubiTaskToRedis.CreateTime
-			}
-
-			if err == nil {
-				ubiTaskRun.Status = constants.UBI_TASK_RUNNING_STATUS
-			} else {
-				ubiTaskRun.Status = constants.UBI_TASK_FAILED_STATUS
-				k8sService := NewK8sService()
-				k8sService.k8sClient.CoreV1().Namespaces().Delete(context.TODO(), namespace, metaV1.DeleteOptions{})
-			}
-			SaveUbiTaskMetadata(ubiTaskRun)
-		}()
-
-		k8sService := NewK8sService()
-		if _, err = k8sService.GetNameSpace(context.TODO(), namespace, metaV1.GetOptions{}); err != nil {
-			if errors.IsNotFound(err) {
-				k8sNamespace := &v1.Namespace{
-					ObjectMeta: metaV1.ObjectMeta{
-						Name: namespace,
-					},
-				}
-				_, err = k8sService.CreateNameSpace(context.TODO(), k8sNamespace, metaV1.CreateOptions{})
-				if err != nil {
-					logs.GetLogger().Errorf("create namespace failed, error: %v", err)
-					return
-				}
-			}
-		}
-
-		receiveUrl := fmt.Sprintf("%s:%d/api/v1/computing/cp/receive/ubi", k8sService.GetAPIServerEndpoint(), conf.GetConfig().API.Port)
-		execCommand := []string{"ubi-bench", "c2"}
-		JobName := strings.ToLower(ubiTask.ZkType) + "-" + strconv.Itoa(ubiTask.ID)
-
-		filC2Param := envVars["FIL_PROOFS_PARAMETER_CACHE"]
-		if gpuFlag == "0" {
-			delete(envVars, "RUST_GPU_TOOLS_CUSTOM_GPU")
-			envVars["BELLMAN_NO_GPU"] = "1"
-		}
-
-		delete(envVars, "FIL_PROOFS_PARAMETER_CACHE")
-		var useEnvVars []v1.EnvVar
-		for k, v := range envVars {
-			useEnvVars = append(useEnvVars, v1.EnvVar{
-				Name:  k,
-				Value: v,
-			})
-		}
-
-		useEnvVars = append(useEnvVars, v1.EnvVar{
-			Name:  "RECEIVE_PROOF_URL",
-			Value: receiveUrl,
-		},
-			v1.EnvVar{
-				Name:  "TASKID",
-				Value: strconv.Itoa(ubiTask.ID),
-			},
-			v1.EnvVar{
-				Name:  "TASK_TYPE",
-				Value: strconv.Itoa(ubiTask.Type),
-			},
-			v1.EnvVar{
-				Name:  "ZK_TYPE",
-				Value: ubiTask.ZkType,
-			},
-			v1.EnvVar{
-				Name:  "NAME_SPACE",
-				Value: namespace,
-			},
-			v1.EnvVar{
-				Name:  "PARAM_URL",
-				Value: ubiTask.InputParam,
-			},
-		)
-
-		job := &batchv1.Job{
-			ObjectMeta: metaV1.ObjectMeta{
-				Name:      JobName,
-				Namespace: namespace,
-			},
-			Spec: batchv1.JobSpec{
-				Template: v1.PodTemplateSpec{
-					Spec: v1.PodSpec{
-						NodeName: nodeName,
-						Containers: []v1.Container{
-							{
-								Name:  JobName + generateString(5),
-								Image: ubiTaskImage,
-								Env:   useEnvVars,
-								VolumeMounts: []v1.VolumeMount{
-									{
-										Name:      "proof-params",
-										MountPath: "/var/tmp/filecoin-proof-parameters",
-									},
-								},
-								Command:         execCommand,
-								Resources:       resourceRequirements,
-								ImagePullPolicy: coreV1.PullIfNotPresent,
-							},
-						},
-						Volumes: []v1.Volume{
-							{
-								Name: "proof-params",
-								VolumeSource: v1.VolumeSource{
-									HostPath: &v1.HostPathVolumeSource{
-										Path: filC2Param,
-									},
-								},
-							},
-						},
-						RestartPolicy: "Never",
-					},
-				},
-				BackoffLimit:            new(int32),
-				TTLSecondsAfterFinished: new(int32),
-			},
-		}
-
-		*job.Spec.BackoffLimit = 1
-		*job.Spec.TTLSecondsAfterFinished = 120
-
-		if _, err = k8sService.k8sClient.BatchV1().Jobs(namespace).Create(context.TODO(), job, metaV1.CreateOptions{}); err != nil {
-			logs.GetLogger().Errorf("Failed creating ubi task job: %v", err)
-			return
-		}
-	}()
-
-	c.JSON(http.StatusOK, util.CreateSuccessResponse("success"))
 }
 
 func ReceiveUbiProof(c *gin.Context) {
-	var c2Proof struct {
-		TaskId    string `json:"task_id"`
-		TaskType  string `json:"task_type"`
-		Proof     string `json:"proof"`
-		ZkType    string `json:"zk_type"`
-		NameSpace string `json:"name_space"`
-	}
-
-	var submitUBIProofTx string
-	var err error
-	defer func() {
-		key := constants.REDIS_UBI_C2_PERFIX + c2Proof.TaskId
-		ubiTask, _ := RetrieveUbiTaskMetadata(key)
-		if err == nil {
-			ubiTask.Status = constants.UBI_TASK_SUCCESS_STATUS
-		} else {
-			ubiTask.Status = constants.UBI_TASK_FAILED_STATUS
-		}
-		ubiTask.Tx = submitUBIProofTx
-		SaveUbiTaskMetadata(ubiTask)
-		if strings.TrimSpace(c2Proof.NameSpace) != "" {
-			k8sService := NewK8sService()
-			k8sService.k8sClient.CoreV1().Namespaces().Delete(context.TODO(), c2Proof.NameSpace, metaV1.DeleteOptions{})
-		}
-	}()
-
+	var c2Proof models.ReceiveProof
 	if err := c.ShouldBindJSON(&c2Proof); err != nil {
 		c.JSON(http.StatusBadRequest, util.CreateErrorResponse(util.JsonError))
 		return
 	}
 	logs.GetLogger().Infof("task_id: %s, C2 proof out received: %+v", c2Proof.TaskId, c2Proof)
 
-	chainUrl, err := conf.GetRpcByName(conf.DefaultRpc)
-	if err != nil {
-		logs.GetLogger().Errorf("get rpc url failed, error: %v,", err)
-		return
+	if constants.GetUBIType(c2Proof.ZkType) == constants.FIL_C2 {
+		Fil_C2_ReceiveUbiProof(c, c2Proof)
+	} else if constants.GetUBIType(c2Proof.ZkType) == constants.ALEO_PROVER {
+		Aleo_Prover_ReceiveUbiProof(c, c2Proof)
 	}
 
-	client, err := ethclient.Dial(chainUrl)
-	if err != nil {
-		logs.GetLogger().Errorf("dial rpc connect failed, error: %v,", err)
-		return
-	}
-	defer client.Close()
-
-	cpStub, err := account.NewAccountStub(client)
-	if err != nil {
-		logs.GetLogger().Errorf("create ubi task client failed, error: %v,", err)
-		return
-	}
-	cpAccount, err := cpStub.GetCpAccountInfo()
-	if err != nil {
-		logs.GetLogger().Errorf("get account info failed, error: %v,", err)
-		return
-	}
-
-	localWallet, err := wallet.SetupWallet(wallet.WalletRepo)
-	if err != nil {
-		logs.GetLogger().Errorf("setup wallet failed, error: %v,", err)
-		return
-	}
-
-	ki, err := localWallet.FindKey(cpAccount.OwnerAddress)
-	if err != nil || ki == nil {
-		logs.GetLogger().Errorf("the address: %s, private key %v,", conf.GetConfig().HUB.WalletAddress, wallet.ErrKeyInfoNotFound)
-		return
-	}
-
-	accountStub, err := account.NewAccountStub(client, account.WithCpPrivateKey(ki.PrivateKey))
-	if err != nil {
-		logs.GetLogger().Errorf("create ubi task client failed, error: %v,", err)
-		return
-	}
-
-	taskType, err := strconv.ParseUint(c2Proof.TaskType, 10, 8)
-	if err != nil {
-		logs.GetLogger().Errorf("conversion to uint8 error: %v", err)
-		return
-	}
-
-	submitUBIProofTx, err = accountStub.SubmitUBIProof(c2Proof.TaskId, uint8(taskType), c2Proof.ZkType, c2Proof.Proof)
-	if err != nil {
-		logs.GetLogger().Errorf("submit ubi proof tx failed, error: %v,", err)
-		return
-	}
-
-	fmt.Printf("submitUBIProofTx: %s", submitUBIProofTx)
-	c.JSON(http.StatusOK, util.CreateSuccessResponse("success"))
 }
 
 func handleConnection(conn *websocket.Conn, spaceDetail models.CacheSpaceDetail, logType string) {
@@ -1479,7 +1141,16 @@ func SaveUbiTaskMetadata(ubiTask *models.CacheUbiTaskDetail) {
 	redisConn := redisPool.Get()
 	defer redisConn.Close()
 
-	key := constants.REDIS_UBI_C2_PERFIX + ubiTask.TaskId
+	var perfix string
+	if constants.GetUBIType(ubiTask.ZkType) == constants.FIL_C2 {
+		perfix = constants.REDIS_UBI_C2_PERFIX
+	}
+
+	if constants.GetUBIType(ubiTask.ZkType) == constants.ALEO_PROVER {
+		perfix = constants.REDIS_UBI_ALEO_PERFIX
+	}
+
+	key := perfix + ubiTask.TaskId
 	redisConn.Do("DEL", redis.Args{}.AddFlat(key)...)
 
 	fullArgs := []interface{}{key}
